java-dest: move the c source files under the native subdirectory
The SyslogNgDestination is splitted into LogDestination base class,
TextLogDestination (for string based messages) and StructuredLogDestination
(for LogMessage based messages).
