+++ modules/java/native/java-plugin.c
