+++ modules/java/native/java_machine.c
