+++ modules/java/native/java_machine.h
